Package Refactoring (missing)
A few missing refactors from the previous commit.
