+++ src/com/islands/games/lifesim/exceptions/TimeException.groovy
@@ -1,7 +1,7 @@
 package com.islands.games.lifesim.exceptions
 
 /**
- * Exception class for issues that arise when calculating {@link com.islands.games.lifesim.Time} object interactions.
+ * Exception class for issues that arise when calculating {@link com.islands.games.lifesim.external.Time} object interactions.
  */
 class TimeException extends Exception {
     TimeException(String msg) {
